MrBiens commit additional code - Product quantity and Type int to Long

+++ src/main/java/com/vn/sbit/SpringMVC/dto/request/purchaseDetail/PurchaseDetailRequest.java
@@ -11,6 +11,6 @@ import lombok.experimental.FieldDefaults;
 public class PurchaseDetailRequest {
     Long purchaseInvoiceId;
     Long productSupplierId;
-    int quantity;
+    Long quantity;
     Double purchasePrice;
 }
